+++ lib/extensions/mod.rs
@@ -6,3 +6,6 @@ pub mod logger;
 
 #[cfg(feature = "serve_static")]
 pub mod serve_static;
+
+#[cfg(feature = "content_type")]
+pub mod content_type;