+++ lib/lib.rs
@@ -9,7 +9,7 @@ Just add the following to your `Cargo.toml`:
 
 ```toml
 [dependencies]
-afire = "0.2.2"
+afire = "0.2.3"
 ```
 
 ## 📄 Info
@@ -60,7 +60,7 @@ For these you will need to enable the feature.
 To use these extra features enable them like this:
 
 ```toml
-afire = { version = "0.2.2", features = ["rate_limit", "logging", "serve_static"] }
+afire = { version = "0.2.3", features = ["rate_limit", "logging", "serve_static"] }
 ```
 
 - Threading
@@ -79,7 +79,7 @@ let mut server: Server = Server::new("localhost", 8080);
 
 #![warn(missing_docs)]
 
-pub(crate) const VERSION: &str = "0.2.2";
+pub(crate) const VERSION: &str = "0.2.3*";
 
 mod common;
 mod http;
@@ -131,3 +131,6 @@ pub use extensions::logger::{Level, Logger};
 
 #[cfg(feature = "serve_static")]
 pub use extensions::serve_static::ServeStatic;
+
+#[cfg(feature = "content_type")]
+pub use extensions::content_type::Content;